requirements.txt
Django>=4.2
requests>=2.31.0
APScheduler>=3.10.4
tzlocal>=5.2
cryptography>=42.0.0
cffi>=1.16.0
sqlparse>=0.4.4
pytz>=2024.1
pycryptodome>=3.20.0
PyYAML>=6.0.2
PyJWT>=2.9.0
